add reservation deleted as same as store

+++ semestralni_prace/server/controllers/v1/mod/stores.js
@@ -5,7 +5,14 @@ import mongoose from 'mongoose';
 import Table from '../../../models/tables.js';
 import { STORES } from '../../../config/stores.js';
 import utils from '../../../utils.js';
-
+import Reservations from '../../../models/reservations.js';
+import dayjs from 'dayjs';
+import utc from 'dayjs/plugin/utc.js';
+import customParseFormat from 'dayjs/plugin/customParseFormat.js';
+import { sendCancellationEmail } from '../../../mailer.js';
+import _ from 'lodash';
+dayjs.extend(utc);
+dayjs.extend(customParseFormat);
 /**
  * @param {import("express").Request} req
  * @param {import("express").Response} res
@@ -92,11 +99,40 @@ export const updateStore = async (req, res, next) => {
 		if (!utils.isValidRequest(addressValidator, req.body.address)) {
 			return next(new HttpError('srv_invalid_request', 400));
 		}
+
+		if (openingTime && (!_.isArray(openingTime) || openingTime.length !== 7)) {
+			return next(new HttpError('srv_invalid_request', 400));
+		}
+
+		openingTime &&
+			openingTime.forEach((day) => {
+				if (
+					!day ||
+					typeof day.isOpen !== 'boolean' ||
+					!day.start ||
+					!dayjs(day.start, 'HH:mm', true).isValid() ||
+					!day.end ||
+					!dayjs(day.end, 'HH:mm', true).isValid() ||
+					dayjs(day.end, 'HH:mm', true).isBefore(
+						dayjs(day.start, 'HH:mm', true)
+					) ||
+					dayjs(day.end, 'HH:mm', true).isSame(dayjs(day.start, 'HH:mm', true))
+				) {
+					console.log('Invalid openingTime');
+					return next(new HttpError('srv_invalid_request', 400));
+				}
+			});
+
+		if (!utils.isValidRequest(addressValidator, req.body.address)) {
+			return next(new HttpError('srv_invalid_request', 400));
+		}
+
 		const validator = {
 			storeId: /^\w{24}$/,
 			name: name ? /^.{3,100}$/ : false,
 			type: type ? utils.createEnumRegex(Object.keys(STORES)) : false,
 			address: false,
+			openingTime: false,
 			isAvailable: isAvailable ? utils.createEnumRegex([true, false]) : false,
 		};
 
@@ -197,13 +233,14 @@ export const addStore = async (req, res, next) => {
  * @param {import("express").NextFunction} next
  */
 export const deleteStore = async (req, res, next) => {
+	let tables = [];
+	let reservations = [];
 	try {
 		const { storeId } = req.params;
 
 		if (!mongoose.Types.ObjectId.isValid(storeId)) {
 			return next(new HttpError('srv_invalid_request', 400));
 		}
-		let tables = [];
 		const store = await Store.findOne({
 			_id: storeId,
 			userId: req.user._id,
@@ -213,9 +250,45 @@ export const deleteStore = async (req, res, next) => {
 			return next(new HttpError('srv_store_not_found', 404));
 		}
 		tables = await Table.find({ storeId: store._id });
+		reservations = await Reservations.find({ storeId: store._id });
 
-		await Table.deleteMany({ storeId: store._id });
+		if (reservations.length > 0) {
+			// Send mail
+			const promises = reservations.map(async (reservation) => {
+				const now = dayjs.utc();
+				//send cancelation email for incoming reservation
+				if (
+					dayjs.utc(reservation.end).isAfter(now) &&
+					!reservation.isCancelled
+				) {
+					if (process.env.NODE_ENV !== 'TEST') {
+						try {
+							const table = tables.find((t) => reservation._id.equals(t._id));
+							await sendCancellationEmail({
+								email: reservation.email,
+								reservation: {
+									...reservation.toObject(),
+									start: dayjs
+										.utc(reservation.start)
+										.format('DD/MM/YYYY HH:mm'),
+									end: dayjs.utc(reservation.end).format('DD/MM/YYYY HH:mm'),
+								},
+								store,
+								table,
+							});
+						} catch (error) {
+							console.log(error);
+						}
+					}
+				}
+			});
+			await Promise.all(promises);
 
+			await Reservations.deleteMany({ storeId: store._id });
+		}
+		if (tables.length > 0) {
+			await Table.deleteMany({ storeId: store._id });
+		}
 		await Store.findByIdAndDelete(store._id);
 
 		return res.status(200).json({
@@ -226,8 +299,9 @@ export const deleteStore = async (req, res, next) => {
 		console.error(error);
 		try {
 			await Table.insertMany(tables);
+			await Reservations.insertMany(reservations);
 		} catch (restoreError) {
-			console.error('Failed to restore tables:', restoreError);
+			console.error('Failed to restore:', restoreError);
 		}
 		return next(new HttpError('srv_delete_store_failed', 500));
 	}
